Type filter working

+++ src/app/filters/type-filter/type-filter.component.ts
@@ -1,20 +1,26 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy, ViewChild, ElementRef } from '@angular/core';
 import { FiltersService } from '../filters.service';
 import { CalendarService } from '../../calendar/calendar.service';
 import { EventType } from '../../shared/event-type.model';
 import { Event } from '../../shared/event.model';
+import { Subscription } from 'rxjs/Subscription';
 
 @Component
 ({
 	selector: 'app-type-filter',
 	templateUrl: './type-filter.component.html'
 })
-export class TypeFilterComponent implements OnInit
+export class TypeFilterComponent implements OnInit, OnDestroy
 {
 	private static filterName = 'type';
 	private static emptyEventType: EventType = new EventType('', '');
 	eventTypes: EventType[] = [];
 	selectedType = '';
+	enabled = false;
+	filtersEnabled = false;
+	enabledSubscription: Subscription;
+	clearedSubscription: Subscription;
+	@ViewChild("types") types: ElementRef;
 	
 	constructor(private filtersService: FiltersService, private calendarService: CalendarService) {}
 	
@@ -25,29 +31,64 @@ export class TypeFilterComponent implements OnInit
 			this.eventTypes = res;
 			this.eventTypes.unshift(TypeFilterComponent.emptyEventType);
 		});
+		
+		this.enabledSubscription = this.filtersService.enabledChanged.subscribe((enabled) =>
+		{
+			this.filtersEnabled = enabled;
+		});
+		
+		this.clearedSubscription = this.filtersService.cleared.subscribe(() =>
+		{
+			this.onClear();
+			this.onEnable(false);
+		});
 	}
 	
 	onChange(eventType: string)
 	{
 		this.selectedType = eventType;
 		
-		if (eventType)
+		if (this.selectedType != TypeFilterComponent.emptyEventType._id)
 		{
 			this.filtersService.addFilter(TypeFilterComponent.filterName, this.filterFunction.bind(this));
 		}
 		else
 		{
-			this.onClear();
+			this.filtersService.removeFilter(TypeFilterComponent.filterName);
+		}
+	}
+	
+	onEnable(enabled: boolean)
+	{
+		this.enabled = enabled;
+		
+		if (this.enabled)
+		{
+			if (this.selectedType)
+			{
+				this.filtersService.addFilter(TypeFilterComponent.filterName, this.filterFunction.bind(this));
+			}
+		}
+		else
+		{
+			this.filtersService.removeFilter(TypeFilterComponent.filterName);
 		}
 	}
 	
 	onClear()
 	{
-		this.filtersService.removeFilter(TypeFilterComponent.filterName);
+		this.onChange(TypeFilterComponent.emptyEventType._id);
+		this.types.nativeElement.value = TypeFilterComponent.emptyEventType._id;
 	}
 	
 	filterFunction(event: Event)
 	{
 		return event.eventCode === this.selectedType;
+	}
+	
+	ngOnDestroy()
+	{
+		this.enabledSubscription.unsubscribe();
+		this.clearedSubscription.unsubscribe();
 	}
 }
